move usenativesearchparams to utils folder, cleanup
