+++ packages/synapse-react-client/src/utils/hooks/index.ts
@@ -17,3 +17,4 @@ export * from './useTraceUpdate'
 export * from './useCookiePreferences'
 export * from '../analytics/useGoogleAnalytics'
 export * from './useSourceAppConfigs'
+export * from './useNativeSearchParams'
